requirements.txt
asn1crypto==0.23.0
bcrypt==3.1.4
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.3
decorator==4.1.2
enum34==1.1.6
Fabric3==1.13.1.post1
idna==2.6
ipaddress==1.0.18
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
paramiko==2.3.1
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.3.7
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.0
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
